requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
